Add a shared syntax file
